speed up database generation by pruning already visited positions
Previously if there are multiple optimal paths to the same state, the
database generation would continue the IDDFS at this position multiple
times, while only continuing once would be enough.
E.g. in the G0->G1 phase it would see that U and U' lead to the same
position, but continue to both UR and U'R.
The updated code remembers for each position the maxDepth value, and if
it the position has been reached during the same maxDepth iteration, it
will skip it, pruning U'R.

This speeds up the database generation by a lot.
Previously the last three database generations all took about a minute,
 now they generate in less than 10 seconds each, G2->G3 even in under a
second.