check if the current branch is optimal before setting the value
With the old approach it was possible that too big values were written
into the database.
E.g. if a position has depth 7, and during a maxDepth=8 search it visits
the same position at depth 8, it would have written 8 into the database
overwriting the 7.

This actually happened regularly. E.g. if you add the following two
lines, you can see that initially it sets the database entry to 7 and
later overwrites it with 8 and 9 during the G1->G2 phase.

  if (depth == maxDepth)
  {
      database.set(index, depth);
+     if (index == 12345)
+         std::cout << "set " << index << " as " << (int)depth << std::endl;
  }